HAI-2039 Seperate hankealue stuff from HankeService to a new service
